requirements.txt
aiogram==3.25.0
SQLAlchemy==2.0.47
aiosqlite==0.22.1
python-dotenv==1.2.1
openpyxl==3.1.5
matplotlib==3.10.8